Example inventory file for openshift-ansible
The example/hosts file can be used as an inventory for
openshift-ansible to install OpenShift. It is probably useless
to the general public, but serves as a working example or an
exhibit.  It requires creating files/htpasswd inside of the
openshift-ansible checkout. The "hosts" file could be copied
to inventory/byo/.